Time.deltaTime

+++ src/rendering/Time.java
@@ -1,5 +1,5 @@
 package rendering;
 
 public class Time {
-static public double deltaTime ; //time since last frame in seconds
+static public float deltaTime ; //time since last frame in seconds
 }
